refactor: move slack and slack-org to lib/zero/
Part of #7697 (restructure src/lib/ into layered architecture).
Move lib/slack/ and lib/slack-org/ to lib/zero/slack/ and
lib/zero/slack-org/ since these modules are zero-layer concerns.

The direct zeroAgents schema import in slack-org/handlers/shared.ts
is now within the zero layer itself, making it architecturally valid.

Closes #7734
