+++ turbo/apps/web/src/lib/org/org-external-cleanup.ts
@@ -5,7 +5,7 @@ import { getStripe } from "../stripe";
 import { deleteWebhook } from "../telegram/client";
 import { decryptSecretValue } from "../shared/crypto/secrets-encryption";
 import { revokeConnectorToken } from "../zero/connector/connector-service";
-import { cleanupWorkspaceInstallation } from "../slack-org/connect-service";
+import { cleanupWorkspaceInstallation } from "../zero/slack-org/connect-service";
 import { orgMetadata } from "../../db/schema/org-metadata";
 import { telegramInstallations } from "../../db/schema/telegram-installation";
 import { agentComposes } from "../../db/schema/agent-compose";
