+++ turbo/apps/web/src/lib/zero/slack-org/handlers/app-home.ts
@@ -1,9 +1,9 @@
 import { eq, and } from "drizzle-orm";
-import { slackOrgInstallations } from "../../../db/schema/slack-org-installation";
-import { slackOrgConnections } from "../../../db/schema/slack-org-connection";
-import { decryptSecretValue } from "../../shared/crypto/secrets-encryption";
-import { env } from "../../../env";
-import { getUserEmail } from "../../auth/get-user-email";
+import { slackOrgInstallations } from "../../../../db/schema/slack-org-installation";
+import { slackOrgConnections } from "../../../../db/schema/slack-org-connection";
+import { decryptSecretValue } from "../../../shared/crypto/secrets-encryption";
+import { env } from "../../../../env";
+import { getUserEmail } from "../../../auth/get-user-email";
 import {
   createSlackClient,
   publishAppHome,
