+++ turbo/apps/web/src/lib/zero/slack-org/handlers/run-agent.ts
@@ -1,11 +1,11 @@
-import { isRunDispatchError } from "../../infra/run";
-import { createZeroRun } from "../../zero/zero-run-service";
-import { buildIntegrationContext } from "../../zero/integration-context";
-import { isApiError } from "../../errors";
+import { isRunDispatchError } from "../../../infra/run";
+import { createZeroRun } from "../../zero-run-service";
+import { buildIntegrationContext } from "../../integration-context";
+import { isApiError } from "../../../errors";
 import { RUN_ERROR_GUIDANCE } from "@vm0/core";
-import { generateCallbackSecret, getApiUrl } from "../../infra/callback";
-import { logger } from "../../logger";
-import type { SlackOrgCallbackPayload } from "../../infra/callback/callback-payloads";
+import { generateCallbackSecret, getApiUrl } from "../../../infra/callback";
+import { logger } from "../../../logger";
+import type { SlackOrgCallbackPayload } from "../../../infra/callback/callback-payloads";
 
 const log = logger("slack-org:run-agent");
