+++ turbo/apps/web/src/lib/zero/slack/resolve-slack-client.ts
@@ -1,11 +1,11 @@
 import type { WebClient } from "@slack/web-api";
 import type { ZeroCapability } from "@vm0/core";
-import { requireAuth, isAuthError } from "../auth/require-auth";
-import { isSandboxAuth } from "../auth/capability-check";
-import { resolveOrg } from "../org/resolve-org";
-import { agentRuns } from "../../db/schema/agent-run";
-import { slackOrgInstallations } from "../../db/schema/slack-org-installation";
-import { decryptSecretValue } from "../shared/crypto/secrets-encryption";
+import { requireAuth, isAuthError } from "../../auth/require-auth";
+import { isSandboxAuth } from "../../auth/capability-check";
+import { resolveOrg } from "../../org/resolve-org";
+import { agentRuns } from "../../../db/schema/agent-run";
+import { slackOrgInstallations } from "../../../db/schema/slack-org-installation";
+import { decryptSecretValue } from "../../shared/crypto/secrets-encryption";
 import { createSlackClient } from "./client";
 import { eq, and } from "drizzle-orm";
